#28 refactor: modify package structure and version upgrade.
- hexagonal architecture -> Separate by package.
- Spring Boot 3.1.4 version upgrade. (lombok issue reason)
- Java 21 version upgrade.
- Gradle 8.5 version upgrade.
- modify root project name to "titi-backend".
- Spring Security 6.1.4 version upgrade and refactor code (method chaining -> lambda)
