+++ settings.gradle
@@ -1,7 +1,1 @@
-rootProject.name = 'backend'
-include 'titi-adapter'
-include 'titi-api'
-include 'titi-application'
-include 'titi-common'
-include 'titi-data'
-include 'titi-domain'
+rootProject.name = 'titi-backend'
