+++ src/main/java/com/titi/auth/adapter/in/security/authentication/AccessDeniedHandlerImpl.java
@@ -1,0 +1,38 @@
+package com.titi.auth.adapter.in.security.authentication;
+
+import static com.titi.common.constant.ResponseCodes.ErrorCode.*;
+
+import java.io.IOException;
+import java.io.OutputStream;
+
+import org.springframework.http.HttpStatus;
+import org.springframework.http.MediaType;
+import org.springframework.security.access.AccessDeniedException;
+import org.springframework.security.web.access.AccessDeniedHandler;
+import org.springframework.stereotype.Component;
+
+import com.fasterxml.jackson.databind.ObjectMapper;
+import jakarta.servlet.http.HttpServletRequest;
+import jakarta.servlet.http.HttpServletResponse;
+import lombok.RequiredArgsConstructor;
+
+import com.titi.common.dto.ErrorResponse;
+
+@Component
+@RequiredArgsConstructor
+public class AccessDeniedHandlerImpl implements AccessDeniedHandler {
+
+	private final ObjectMapper objectMapper;
+
+	@Override
+	public void handle(HttpServletRequest request, HttpServletResponse response, AccessDeniedException accessDeniedException) throws IOException {
+		response.setStatus(HttpStatus.FORBIDDEN.value());
+		response.setContentType(MediaType.APPLICATION_JSON_VALUE);
+
+		try (OutputStream os = response.getOutputStream()) {
+			objectMapper.writeValue(os, ErrorResponse.of(INSUFFICIENT_AUTHORITY));
+			os.flush();
+		}
+	}
+
+}
