+++ src/main/java/com/titi/common/constant/Constants.java
@@ -1,0 +1,13 @@
+package com.titi.common.constant;
+
+import lombok.AccessLevel;
+import lombok.NoArgsConstructor;
+
+@NoArgsConstructor(access = AccessLevel.PRIVATE)
+public final class Constants {
+
+	public static final String EMPTY = "";
+	public static final String SINGLE_SPACE = " ";
+	public static final String DOT = ".";
+
+}
