+++ src/main/java/com/titi/common/dto/ErrorResponse.java
@@ -1,0 +1,70 @@
+package com.titi.common.dto;
+
+import java.util.ArrayList;
+import java.util.List;
+
+import lombok.AccessLevel;
+import lombok.Getter;
+import lombok.NoArgsConstructor;
+
+import com.titi.common.constant.ResponseCodes.ErrorCode;
+
+@Getter
+public class ErrorResponse {
+
+	private final int status;
+	private final String code;
+	private final String message;
+	private final List<FieldError> errors;
+
+	private ErrorResponse(ErrorCode errorCode, List<FieldError> errors) {
+		this.status = errorCode.getStatus();
+		this.code = errorCode.getCode();
+		this.message = errorCode.getMessage();
+		this.errors = errors;
+	}
+
+	private ErrorResponse(ErrorCode errorCode) {
+		this.status = errorCode.getStatus();
+		this.code = errorCode.getCode();
+		this.message = errorCode.getMessage();
+		this.errors = new ArrayList<>();
+	}
+
+	public static ErrorResponse of(final ErrorCode errorCode) {
+		return new ErrorResponse(errorCode);
+	}
+
+	public static ErrorResponse of(final ErrorCode errorCode, final List<FieldError> errors) {
+		return new ErrorResponse(errorCode, errors);
+	}
+
+	@Getter
+	@NoArgsConstructor(access = AccessLevel.PROTECTED)
+	public static class FieldError {
+
+		private String field;
+		private String value;
+		private String reason;
+
+		public FieldError(String field, String value, String reason) {
+			this.field = field;
+			this.value = value;
+			this.reason = reason;
+		}
+
+		public static List<FieldError> of(String field, String value, String reason) {
+			final List<FieldError> fieldErrors = new ArrayList<>();
+			fieldErrors.add(new FieldError(field, value, reason));
+			return fieldErrors;
+		}
+
+		public static List<FieldError> of(String field, String value, ErrorCode errorCode) {
+			final List<FieldError> fieldErrors = new ArrayList<>();
+			fieldErrors.add(new FieldError(field, value, errorCode.getMessage()));
+			return fieldErrors;
+		}
+
+	}
+
+}
