+++ src/main/java/com/titi/common/exception/TiTiException.java
@@ -1,0 +1,30 @@
+package com.titi.common.exception;
+
+import java.util.ArrayList;
+import java.util.List;
+
+
+import lombok.Getter;
+
+import com.titi.common.constant.ResponseCodes;
+import com.titi.common.dto.ErrorResponse.FieldError;
+
+@Getter
+public class TiTiException extends RuntimeException {
+
+	private final ResponseCodes.ErrorCode errorCode;
+	private final List<FieldError> errors;
+
+	public TiTiException(ResponseCodes.ErrorCode errorCode, List<FieldError> errors) {
+		super(errorCode.getMessage());
+		this.errorCode = errorCode;
+		this.errors = errors;
+	}
+
+	public TiTiException(ResponseCodes.ErrorCode errorCode) {
+		super(errorCode.getMessage());
+		this.errorCode = errorCode;
+		this.errors = new ArrayList<>();
+	}
+
+}
