--- titi-adapter/build.gradle
@@ -1,9 +1,0 @@
-dependencies {
-    api project(":titi-data")
-    api project(":titi-application")
-
-    testImplementation("org.springframework.boot:spring-boot-starter-test")
-}
-
-tasks.bootJar { enabled = false }
-tasks.jar { enabled = true }
