--- titi-api/src/main/java/com/titi/security/authentication/AccessDeniedHandlerImpl.java
@@ -1,37 +1,0 @@
-package com.titi.security.authentication;
-
-import java.io.IOException;
-import java.io.OutputStream;
-
-import org.springframework.http.HttpStatus;
-import org.springframework.http.MediaType;
-import org.springframework.security.access.AccessDeniedException;
-import org.springframework.security.web.access.AccessDeniedHandler;
-import org.springframework.stereotype.Component;
-
-import com.fasterxml.jackson.databind.ObjectMapper;
-import jakarta.servlet.http.HttpServletRequest;
-import jakarta.servlet.http.HttpServletResponse;
-import lombok.RequiredArgsConstructor;
-
-import com.titi.common.ResponseCodes.ErrorCode;
-import com.titi.exception.ErrorResponse;
-
-@Component
-@RequiredArgsConstructor
-public class AccessDeniedHandlerImpl implements AccessDeniedHandler {
-
-	private final ObjectMapper objectMapper;
-
-	@Override
-	public void handle(HttpServletRequest request, HttpServletResponse response, AccessDeniedException accessDeniedException) throws IOException {
-		response.setStatus(HttpStatus.FORBIDDEN.value());
-		response.setContentType(MediaType.APPLICATION_JSON_VALUE);
-
-		try (OutputStream os = response.getOutputStream()) {
-			objectMapper.writeValue(os, ErrorResponse.of(ErrorCode.INSUFFICIENT_AUTHORITY));
-			os.flush();
-		}
-	}
-
-}
