--- titi-api/src/main/resources/application.yml
@@ -1,6 +1,0 @@
-server:
-  port: 8080
-  shutdown: graceful
-
-jwt:
-  secret-key: abcdefghijklmnopqrstuwxzy0123456
