--- titi-application/build.gradle
@@ -1,13 +1,0 @@
-dependencies {
-    api project(":titi-common")
-    api project(":titi-domain")
-
-    implementation 'org.springframework.boot:spring-boot-starter-web'
-
-    implementation 'io.jsonwebtoken:jjwt-api:0.12.3'
-    runtimeOnly 'io.jsonwebtoken:jjwt-impl:0.12.3'
-    runtimeOnly 'io.jsonwebtoken:jjwt-jackson:0.12.3'
-}
-
-tasks.bootJar { enabled = false }
-tasks.jar { enabled = true }
