--- titi-common/build.gradle
@@ -1,2 +1,0 @@
-tasks.bootJar { enabled = false }
-tasks.jar { enabled = true }
