--- titi-common/src/main/java/com/titi/common/Constants.java
@@ -1,13 +1,0 @@
-package com.titi.common;
-
-import lombok.AccessLevel;
-import lombok.NoArgsConstructor;
-
-@NoArgsConstructor(access = AccessLevel.PRIVATE)
-public final class Constants {
-
-	public static final String EMPTY = "";
-	public static final String SINGLE_SPACE = " ";
-	public static final String DOT = ".";
-
-}
