--- titi-data/build.gradle
@@ -1,13 +1,0 @@
-dependencies {
-    implementation project(":titi-common")
-    implementation project(":titi-domain")
-
-    api 'org.springframework.boot:spring-boot-starter-data-jpa'
-    runtimeOnly 'com.mysql:mysql-connector-j'
-    runtimeOnly 'com.h2database:h2'
-
-    testImplementation("org.springframework.boot:spring-boot-starter-test")
-}
-
-tasks.bootJar { enabled = false }
-tasks.jar { enabled = true }
