--- titi-domain/build.gradle
@@ -1,6 +1,0 @@
-dependencies {
-    implementation project(":titi-common")
-}
-
-tasks.bootJar { enabled = false }
-tasks.jar { enabled = true }
